scripts/train_model.py
#!/usr/bin/env python3
"""
ECG-LLM Model Training Script

This script trains the ECG-LLM model for cardiovascular diagnosis.

Usage:
    python scripts/train_model.py --config config/config.yaml --experiment baseline
"""

import argparse
import yaml
import os
from pathlib import Path
import sys

# Add src to path
sys.path.append(str(Path(__file__).parent.parent / 'src'))

from utils.logger import ECGLogger
from utils.data_loader import ECGDataLoader
from preprocessing.signal_filter import ECGFilter
from encoding.symbolic_encoder import SymbolicEncoder
from encoding.template_generator import TemplateGenerator
from models.ecg_llm import ECGLLM
from evaluation.metrics import ECGMetrics


def parse_args():
    """Parse command line arguments."""
    parser = argparse.ArgumentParser(description='Train ECG-LLM model')
    parser.add_argument('--config', type=str, default='config/config.yaml',
                       help='Path to configuration file')
    parser.add_argument('--experiment', type=str, default='baseline',
                       help='Experiment name')
    parser.add_argument('--data-path', type=str, default='data/raw/',
                       help='Path to training data')
    parser.add_argument('--output-dir', type=str, default='experiments/',
                       help='Output directory for results')
    parser.add_argument('--verbose', action='store_true',
                       help='Enable verbose logging')
    
    return parser.parse_args()


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def setup_experiment(config: dict, experiment_name: str, output_dir: str):
    """Set up experiment directory and logging."""
    experiment_dir = Path(output_dir) / experiment_name
    experiment_dir.mkdir(parents=True, exist_ok=True)
    
    # Save configuration
    config_file = experiment_dir / 'config.yaml'
    with open(config_file, 'w') as f:
        yaml.dump(config, f, default_flow_style=False)
    
    # Set up logger
    logger = ECGLogger(f'train_{experiment_name}', experiment_name)
    logger.log_experiment_start(config)
    
    return experiment_dir, logger


def load_and_preprocess_data(config: dict, data_path: str, logger: ECGLogger):
    """Load and preprocess ECG data."""
    logger.logger.info("Loading ECG data...")
    
    # Initialize data loader
    loader = ECGDataLoader(data_path)
    
    # Load data based on dataset type
    dataset_name = config['data']['datasets'][0]['name']
    
    if dataset_name == 'PTB-XL':
        signals, metadata = loader.load_ptb_xl()
    elif dataset_name == 'MIT-BIH':
        signals, metadata = loader.load_mit_bih()
    else:
        signals, metadata = loader.load_custom_dataset()
    
    logger.log_data_info(dataset_name, len(signals), signals.shape[1] if signals.ndim > 1 else 1)
    
    # Preprocess signals
    logger.logger.info("Preprocessing ECG signals...")
    preprocessor = ECGFilter(sampling_rate=config['data']['datasets'][0]['sampling_rate'])
    
    preprocessed_signals = []
    for signal in signals:
        processed = preprocessor.preprocess_ecg(
            signal,
            apply_bandpass=config['preprocessing']['filter'].get('apply_bandpass', True),
            apply_notch=config['preprocessing']['filter'].get('apply_notch', True),
            apply_baseline=config['preprocessing']['filter'].get('apply_baseline', True)
        )
        preprocessed_signals.append(processed)
    
    return preprocessed_signals, metadata


def encode_signals(signals: list, config: dict, logger: ECGLogger):
    """Encode ECG signals to text representations."""
    logger.logger.info("Encoding ECG signals to text...")
    
    encoding_config = config['encoding']
    
    if encoding_config.get('use_symbolic', True):
        # Symbolic encoding
        encoder = SymbolicEncoder(
            vocab_size=encoding_config['symbolic']['vocab_size'],
            segment_length=encoding_config['symbolic']['segment_tokens']
        )
        
        # Fit encoder on training data
        encoder.fit(signals)
        
        # Encode signals
        encoded_signals = encoder.encode_batch(signals)
        
        # Convert to text
        ecg_texts = [' '.join(tokens) for tokens in encoded_signals]
        
    else:
        # Template-based encoding
        template_generator = TemplateGenerator(
            use_medical_terms=encoding_config['template']['use_medical_terms'],
            include_measurements=encoding_config['template']['include_measurements']
        )
        
        ecg_texts = [template_generator.generate(signal) for signal in signals]
    
    logger.logger.info(f"Encoded {len(ecg_texts)} ECG signals")
    return ecg_texts


def train_model(ecg_texts: list, metadata: dict, config: dict, logger: ECGLogger):
    """Train the ECG-LLM model."""
    logger.logger.info("Initializing ECG-LLM model...")
    
    model_config = config['model']
    
    # Initialize model
    model = ECGLLM(
        model_name=model_config['base_model'],
        config_path=None  # Use default config
    )
    
    logger.log_model_info(model_config['base_model'], 0)  # Placeholder for parameter count
    
    # For demonstration, we'll simulate training
    # In practice, you would implement actual fine-tuning here
    
    logger.logger.info("Training model (simulation)...")
    
    # Simulate training epochs
    for epoch in range(model_config['fine_tuning'].get('epochs', 3)):
        # Simulate batch processing
        batch_size = model_config['fine_tuning'].get('batch_size', 16)
        
        for i in range(0, len(ecg_texts), batch_size):
            batch_texts = ecg_texts[i:i+batch_size]
            
            # Simulate training step
            # In practice, you would:
            # 1. Create training prompts
            # 2. Get model predictions
            # 3. Calculate loss
            # 4. Update model parameters
            
            pass
        
        # Log simulated metrics
        simulated_metrics = {
            'loss': 0.5 - epoch * 0.1,
            'accuracy': 0.7 + epoch * 0.1,
            'f1_score': 0.65 + epoch * 0.12
        }
        
        logger.log_metrics(simulated_metrics, epoch)
    
    logger.logger.info("Model training completed")
    return model


def evaluate_model(model: ECGLLM, ecg_texts: list, metadata: dict, config: dict, logger: ECGLogger):
    """Evaluate the trained model."""
    logger.logger.info("Evaluating model...")
    
    # Initialize metrics
    metrics = ECGMetrics()
    
    # Simulate evaluation
    predictions = []
    true_labels = []
    
    for i, ecg_text in enumerate(ecg_texts[:10]):  # Evaluate on first 10 samples
        # Get model prediction
        try:
            result = model.diagnose(
                ecg_text=ecg_text,
                patient_info={'age': 'unknown', 'gender': 'unknown', 'symptoms': 'none'},
                include_explanation=False
            )
            predictions.append(result['primary_diagnosis'])
            
            # Simulate true label
            true_labels.append('Normal')  # Placeholder
            
        except Exception as e:
            logger.logger.error(f"Evaluation error for sample {i}: {e}")
            predictions.append('Error')
            true_labels.append('Unknown')
    
    # Calculate metrics (simplified)
    accuracy = sum(1 for p, t in zip(predictions, true_labels) if p == t) / len(predictions)
    
    evaluation_metrics = {
        'accuracy': accuracy,
        'predictions': len(predictions),
        'errors': predictions.count('Error')
    }
    
    logger.log_metrics(evaluation_metrics)
    
    return evaluation_metrics


def save_results(model: ECGLLM, metrics: dict, experiment_dir: Path, logger: ECGLogger):
    """Save model and results."""
    logger.logger.info("Saving results...")
    
    # Save metrics
    metrics_file = experiment_dir / 'metrics.yaml'
    with open(metrics_file, 'w') as f:
        yaml.dump(metrics, f, default_flow_style=False)
    
    # Save model info
    model_info = model.get_model_info()
    model_info_file = experiment_dir / 'model_info.yaml'
    with open(model_info_file, 'w') as f:
        yaml.dump(model_info, f, default_flow_style=False)
    
    logger.logger.info(f"Results saved to {experiment_dir}")


def main():
    """Main training function."""
    args = parse_args()
    
    # Load configuration
    config = load_config(args.config)
    
    # Set up experiment
    experiment_dir, logger = setup_experiment(config, args.experiment, args.output_dir)
    
    try:
        # Load and preprocess data
        signals, metadata = load_and_preprocess_data(config, args.data_path, logger)
        
        # Encode signals to text
        ecg_texts = encode_signals(signals, config, logger)
        
        # Train model
        model = train_model(ecg_texts, metadata, config, logger)
        
        # Evaluate model
        evaluation_metrics = evaluate_model(model, ecg_texts, metadata, config, logger)
        
        # Save results
        save_results(model, evaluation_metrics, experiment_dir, logger)
        
        logger.log_experiment_end()
        
        print(f"Training completed successfully! Results saved to {experiment_dir}")
        
    except Exception as e:
        logger.logger.error(f"Training failed: {e}")
        raise


if __name__ == '__main__':
    main()