requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Signal processing
wfdb==4.1.0
neurokit2==0.2.4
pywt==1.4.1
scikit-learn==1.3.0

# Deep learning
torch==2.0.1
transformers==4.32.0
datasets==2.14.0
accelerate==0.21.0
peft==0.4.0

# LLM APIs
openai==0.27.8
anthropic==0.3.11
google-generativeai==0.1.0

# RAG and knowledge graphs
langchain==0.0.229
faiss-cpu==1.7.4
chromadb==0.4.6
networkx==3.1

# Medical data
pydicom==2.4.3
nibabel==5.1.0
medpy==0.4.0

# Data processing
tqdm==4.65.0
joblib==1.3.2
dask==2023.7.1

# Evaluation
sklearn-metrics==0.0.4
wandb==0.15.8

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.5.2

# Development
jupyter==1.0.0
ipykernel==6.25.0
black==23.7.0
pytest==7.4.0
mypy==1.5.1
flake8==6.0.0

# Optional: For CUDA support
# torch-audio==2.0.1
# torch-vision==0.15.1